Avoid panic on u16 << 16

+++ src/neon/y_p16_to_rgba16.rs
@@ -52,8 +52,8 @@ pub(crate) unsafe fn neon_y_p16_to_rgba16_row<
 
     let y_corr = vdupq_n_u16(range.bias_y as u16);
     let v_luma_coeff = vdupq_n_u16(transform.y_coef as u16);
-    let v_alpha = vdupq_n_u16((1 << BIT_DEPTH) - 1);
-    let v_max_values = vdupq_n_u16((1 << BIT_DEPTH) - 1);
+    let v_alpha = vdupq_n_u16(((1u32 << BIT_DEPTH) - 1) as u16);
+    let v_max_values = vdupq_n_u16(((1u32 << BIT_DEPTH) - 1) as u16);
     let rnd_base = vdupq_n_u32(1 << (PRECISION - 1));
 
     let mut cx = start_cx;
